Move CheckExplainConfig to a sub-package and fix pmd errors
